+++ backend/schedulerd/cron_scheduler.go
@@ -6,7 +6,7 @@ import (
 	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/backend/secrets"
 	"github.com/sensu/sensu-go/backend/store"
-	cachev3 "github.com/sensu/sensu-go/backend/store/cache/v3"
+	cachev2 "github.com/sensu/sensu-go/backend/store/cache/v2"
 	"github.com/sirupsen/logrus"
 
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
@@ -22,12 +22,12 @@ type CronScheduler struct {
 	ctx                    context.Context
 	cancel                 context.CancelFunc
 	interrupt              chan *corev2.CheckConfig
-	entityCache            *cachev3.Resource
+	entityCache            *cachev2.Resource
 	secretsProviderManager *secrets.ProviderManager
 }
 
 // NewCronScheduler initializes a CronScheduler
-func NewCronScheduler(ctx context.Context, store store.Store, bus messaging.MessageBus, check *corev2.CheckConfig, cache *cachev3.Resource, secretsProviderManager *secrets.ProviderManager) *CronScheduler {
+func NewCronScheduler(ctx context.Context, store store.Store, bus messaging.MessageBus, check *corev2.CheckConfig, cache *cachev2.Resource, secretsProviderManager *secrets.ProviderManager) *CronScheduler {
 	sched := &CronScheduler{
 		store:         store,
 		bus:           bus,
