+++ backend/schedulerd/proxy_check.go
@@ -8,7 +8,7 @@ import (
 	cron "github.com/robfig/cron/v3"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	corev3 "github.com/sensu/sensu-go/api/core/v3"
-	cachev3 "github.com/sensu/sensu-go/backend/store/cache/v3"
+	cachev2 "github.com/sensu/sensu-go/backend/store/cache/v2"
 	"github.com/sensu/sensu-go/js"
 	"github.com/sensu/sensu-go/token"
 	"github.com/sensu/sensu-go/types/dynamic"
@@ -16,7 +16,7 @@ import (
 
 // matchEntities matches the provided list of entities to the entity attributes
 // configured in the proxy request
-func matchEntities(entities []cachev3.Value, proxyRequest *corev2.ProxyRequests) []*corev3.EntityConfig {
+func matchEntities(entities []cachev2.Value, proxyRequest *corev2.ProxyRequests) []*corev3.EntityConfig {
 	matched := make([]*corev3.EntityConfig, 0, len(entities))
 	synthesizedEntities := make([]interface{}, 0, len(entities))
 	for _, entity := range entities {
