+++ backend/store/cache/cache_test.go
@@ -1,30 +1,31 @@
-package v3
+package cache
 
 import (
 	"context"
 	"reflect"
 	"testing"
 
+	"github.com/sensu/sensu-go/backend/store/etcd"
+	"github.com/sensu/sensu-go/types"
+
 	"github.com/coreos/etcd/integration"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
-	corev3 "github.com/sensu/sensu-go/api/core/v3"
 	"github.com/sensu/sensu-go/backend/store"
-	storev2 "github.com/sensu/sensu-go/backend/store/v2"
-	"github.com/sensu/sensu-go/backend/store/v2/etcdstore"
-	"github.com/sensu/sensu-go/backend/store/v2/wrap"
+	"github.com/sensu/sensu-go/testing/fixture"
 	"github.com/sensu/sensu-go/types/dynamic"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func fixtureEntity(namespace, name string) *corev3.EntityConfig {
-	entity := corev3.FixtureEntityConfig(name)
-	entity.Metadata.Namespace = namespace
+func fixtureEntity(namespace, name string) *corev2.Entity {
+	entity := corev2.FixtureEntity(name)
+	entity.Namespace = namespace
 	return entity
 }
 
 func TestCacheGet(t *testing.T) {
 	cache := Resource{
-		cache: buildCache([]corev3.Resource{
+		cache: buildCache([]corev2.Resource{
 			fixtureEntity("a", "1"),
 			fixtureEntity("a", "2"),
 			fixtureEntity("a", "3"),
@@ -63,7 +64,7 @@ func TestCacheGet(t *testing.T) {
 
 func TestCacheGetAll(t *testing.T) {
 	cache := Resource{
-		cache: buildCache([]corev3.Resource{
+		cache: buildCache([]corev2.Resource{
 			fixtureEntity("a", "1"),
 			fixtureEntity("a", "2"),
 			fixtureEntity("b", "1"),
@@ -96,12 +97,11 @@ func TestCacheGetAll(t *testing.T) {
 }
 
 func TestBuildCache(t *testing.T) {
-	resource1 := corev3.FixtureEntityConfig("resource1")
-	resource2 := corev3.FixtureEntityConfig("resource2")
-	resource3 := corev3.FixtureEntityConfig("resource3")
-	resource3.Metadata.Namespace = "acme"
+	resource1 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource1", Namespace: "default"}}
+	resource2 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource2", Namespace: "default"}}
+	resource3 := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "resource3", Namespace: "acme"}}
 
-	cache := buildCache([]corev3.Resource{resource1, resource2, resource3}, false)
+	cache := buildCache([]corev2.Resource{resource1, resource2, resource3}, false)
 
 	assert.Len(t, cache["acme"], 1)
 	assert.Len(t, cache["default"], 2)
@@ -109,37 +109,22 @@ func TestBuildCache(t *testing.T) {
 }
 
 func TestResourceRebuild(t *testing.T) {
-	ctx := store.NamespaceContext(context.Background(), "default")
 	c := integration.NewClusterV3(t, &integration.ClusterConfig{Size: 1})
 	defer c.Terminate(t)
 	client := c.RandClient()
-	store := etcdstore.NewStore(client)
-
-	// Add namespace resource
-	namespace := corev2.FixtureNamespace("default")
-	req := storev2.NewResourceRequestFromV2Resource(ctx, namespace)
-	wrapper, err := wrap.V2Resource(namespace)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := store.CreateOrUpdate(req, wrapper); err != nil {
-		t.Fatal(err)
-	}
+	s := etcd.NewStore(client, "store")
+	require.NoError(t, s.CreateNamespace(context.Background(), types.FixtureNamespace("default")))
+	ctx := store.NamespaceContext(context.Background(), "default")
 
 	cacher := Resource{
 		cache:     make(map[string][]Value),
 		client:    client,
-		resourceT: &corev3.EntityConfig{},
+		resourceT: &fixture.Resource{},
 	}
 
 	// Resource added to a new namespace
-	foo := corev3.FixtureEntityConfig("foo")
-	req = storev2.NewResourceRequestFromResource(ctx, foo)
-	wrapper, err = wrap.Resource(foo)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := store.CreateOrUpdate(req, wrapper); err != nil {
+	foo := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "foo", Namespace: "default"}}
+	if err := s.CreateOrUpdateResource(ctx, foo); err != nil {
 		t.Fatal(err)
 	}
 	if updates, err := cacher.rebuild(ctx); err != nil {
@@ -151,13 +136,8 @@ func TestResourceRebuild(t *testing.T) {
 	assert.Equal(t, int64(1), cacher.Count())
 
 	// Resource added to an existing namespace
-	bar := corev3.FixtureEntityConfig("bar")
-	req = storev2.NewResourceRequestFromResource(ctx, bar)
-	wrapper, err = wrap.Resource(bar)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := store.CreateOrUpdate(req, wrapper); err != nil {
+	bar := &fixture.Resource{ObjectMeta: corev2.ObjectMeta{Name: "bar", Namespace: "default"}}
+	if err := s.CreateOrUpdateResource(ctx, bar); err != nil {
 		t.Fatal(err)
 	}
 	if updates, err := cacher.rebuild(ctx); err != nil {
@@ -169,13 +149,8 @@ func TestResourceRebuild(t *testing.T) {
 	assert.Equal(t, int64(2), cacher.Count())
 
 	// Resource updated
-	bar.User = "acme"
-	req = storev2.NewResourceRequestFromResource(ctx, bar)
-	wrapper, err = wrap.Resource(bar)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := store.CreateOrUpdate(req, wrapper); err != nil {
+	bar.Foo = "acme"
+	if err := s.CreateOrUpdateResource(ctx, bar); err != nil {
 		t.Fatal(err)
 	}
 	if updates, err := cacher.rebuild(ctx); err != nil {
@@ -187,8 +162,7 @@ func TestResourceRebuild(t *testing.T) {
 	assert.Equal(t, int64(2), cacher.Count())
 
 	// Resource deleted
-	req = storev2.NewResourceRequestFromResource(ctx, bar)
-	if err := store.Delete(req); err != nil {
+	if err := s.DeleteResource(ctx, bar.StorePrefix(), bar.GetObjectMeta().Name); err != nil {
 		t.Fatal(err)
 	}
 	if updates, err := cacher.rebuild(ctx); err != nil {
