+++ backend/store/cache/logger.go
@@ -1,8 +1,8 @@
-package v3
+package cache
 
 import "github.com/sirupsen/logrus"
 
 var logger = logrus.WithFields(logrus.Fields{
 	"component":   "cache",
-	"api_version": "v3",
+	"api_version": "v1",
 })
